Update CONTRIBUTING and README files for clarity
Enhanced the `CONTRIBUTING.md` file with a welcoming introduction, a detailed table of contents, and structured sections to guide potential contributors.

Modified the `README.md` to improve the library description, update links to the GitHub repository and documentation, and direct users to the new `CONTRIBUTING.md`.

Changed author attribution in `README.md` to "Made with ❤️ by Mohamed OUBAICH".